feat: Add husky and lint-staged for formatting
