+++ package.json
@@ -4,8 +4,11 @@
   "main": "server.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon server.js",
-    "start": "node server.js"
+    "dev": "nodemon src/server.js",
+    "start": "node src/server.js",
+    "prepare": "husky",
+    "check-format": "prettier --check ./src",
+    "format": "prettier --write ./src"
   },
   "keywords": [],
   "author": "",
@@ -27,9 +30,17 @@
     "node-cron": "^3.0.3"
   },
   "devDependencies": {
-    "nodemon": "^3.1.9"
+    "husky": "^9.1.7",
+    "lint-staged": "^16.1.0",
+    "nodemon": "^3.1.9",
+    "prettier": "^3.5.3"
   },
   "engines": {
     "node": "18.x"
+  },
+  "lint-staged": {
+    "src/**/*.{ts,js,md,json}": [
+      "npm run check-format"
+    ]
   }
 }
